third_party/blink/web_tests/svg/as-image/svg-canvas-svg-with-image-with-link-tainted-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS window.ctx.getImageData(0, 0, 1, 1) threw exception SecurityError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The canvas has been tainted by cross-origin data..
